requirements.txt
absl-py==1.2.0
altair==4.1.0
argon2-cffi==21.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
boto==2.49.0
Brotli==1.0.9
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
cryptography==36.0.1
cssselect==1.1.0
cycler==0.11.0
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
docopt==0.6.2
entrypoints==0.3
execnet==1.9.0
fasteners==0.17.2
filelock==3.4.2
Flask==2.0.2
Flask-Compress==1.10.1
flatbuffers==22.9.24
fonttools==4.28.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.24
google==3.0.0
google-api-core==2.4.0
google-api-python-client==2.68.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.54.0
grpcio==1.49.1
h11==0.13.0
h5py==3.7.0
httplib2==0.20.2
idna==3.3
importlib-metadata==5.1.0
iniconfig==1.1.1
ipdb==0.13.9
ipykernel==6.6.0
ipython==7.31.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.2.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
more-itertools==8.12.0
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.3
nose==1.3.7
notebook==6.4.6
numpy==1.21.4
oauth2client==4.1.3
oauthlib==3.2.1
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parameterized==0.8.1
parso==0.8.3
patsy==0.5.2
pdfminer.six==20211012
pep517==0.13.0
pickleshare==0.7.5
Pillow==9.0.0
pipreqs==0.4.11
platformdirs==2.4.1
plotly==5.4.0
plotly-express==0.4.1
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.1
py==1.11.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.7.1
Pygments==2.11.2
Pympler==0.9
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyotp==2.6.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
pytest-forked==1.4.0
pytest-html==2.0.1
pytest-metadata==1.11.0
pytest-ordering==0.6
pytest-rerunfailures==10.2
pytest-xdist==2.5.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==302
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rich==11.0.0
rsa==4.8
sbvirtualdisplay==1.0.0
scipy==1.7.3
seaborn==0.12.1
selenium==4.1.0
seleniumbase==2.4.0
semver==2.13.0
Send2Trash==1.8.0
setuptools-scm==6.4.2
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
statsmodels==0.13.1
streamlit==1.15.1
streamlit-lottie==0.0.3
streamlit-option-menu==0.3.2
tenacity==8.0.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tomli==2.0.0
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2021.5
tzlocal==4.1
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.8
vaderSentiment==3.3.2
validators==0.18.2
virtualenv==20.13.0
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.14.1
wsproto==1.0.0
yarg==0.1.9
zipp==3.11.0
